refactor: extract item frame interaction related events to independent listener
